from https://github.com/lukfugl/fake_sqs/tree/long-polling
support for long polling (WaitTimeSeconds)

allow a ReceiveMessage action to know whether an attempt (via call)
satisified the request, and then have the dispatcher in the API loop
until it is satisfied, sleeping in between, rather than just making one
attempt.

the ReceiveMessage action is satisfied by either return 1+ results, or
by expiring. it expires according to the action's WaitTimeSeconds or the
queue's ReceiveMessageWaitTimeSeconds. if neither is set, or is set to
0, then only one attempt is made before immediately expiring.